Move RequestReader, ResponseSender and ReadRequestException to server package
