+++ src/main/java/httpserver/server/ExceptionHandler.java
@@ -1,7 +1,5 @@
 package httpserver.server;
 
-import httpserver.http.request.ReadRequestException;
-
 
 public class ExceptionHandler {
